+++ packages/aws-client/StepFunctions.js
@@ -193,6 +193,7 @@ const unicodeEscape = (str, regex = /[\s\S]/g) => str.replace(regex, unicodeEsca
  * @returns {string} A string that's safe to use as a StepFunctions execution name
  */
 const toSfnExecutionName = (fields, delimiter = '__') => {
+  deprecate('@cumulus/aws-client/StepFunctions.toSfnExecutionName()', '1.20.0');
   let sfnUnsafeChars = '[^\\w-=+_.]';
   if (delimiter) {
     sfnUnsafeChars = `(${delimiter}|${sfnUnsafeChars})`;
@@ -214,10 +215,12 @@ const toSfnExecutionName = (fields, delimiter = '__') => {
  *   no replacements
  * @returns {Array} An array of the original fields
  */
-const fromSfnExecutionName = (str, delimiter = '__') =>
-  str.split(delimiter)
+const fromSfnExecutionName = (str, delimiter = '__') => {
+  deprecate('@cumulus/aws-client/StepFunctions.fromSfnExecutionName()', '1.20.0');
+  return str.split(delimiter)
     .map((s) => s.replace(/!/g, '\\').replace('"', '\\"'))
     .map((s) => JSON.parse(`"${s}"`));
+}
 
 /**
  * Returns execution ARN from a statement machine Arn and executionName
@@ -227,6 +230,7 @@ const fromSfnExecutionName = (str, delimiter = '__') =>
  * @returns {string} - Step Function Execution Arn
  */
 const getExecutionArn = (stateMachineArn, executionName) => {
+  deprecate('@cumulus/aws-client/StepFunctions.getExecutionArn()', '1.20.0', '@cumulus/message/executions.buildExecutionArn()');
   if (stateMachineArn && executionName) {
     const sfArn = stateMachineArn.replace('stateMachine', 'execution');
     return `${sfArn}:${executionName}`;
@@ -241,12 +245,14 @@ const getExecutionArn = (stateMachineArn, executionName) => {
  * @returns {string} return aws console url for the execution
  */
 function getExecutionUrl(executionArn) {
+  deprecate('@cumulus/aws-client/StepFunctions.getExecutionUrl()', '1.20.0', '@cumulus/message/executions.getExecutionUrlFromArn()');
   const region = process.env.AWS_DEFAULT_REGION || 'us-east-1';
   return `https://console.aws.amazon.com/states/home?region=${region}`
          + `#/executions/details/${executionArn}`;
 }
 
 const getStateMachineArn = (executionArn) => {
+  deprecate('@cumulus/aws-client/StepFunctions.getStateMachineArn()', '1.20.0', '@cumulus/message/executions.getStateMachineArnFromExecutionArn()');
   if (executionArn) {
     return executionArn.replace('execution', 'stateMachine').split(':').slice(0, -1).join(':');
   }